python/dist-langchain/examples/basic_usage/requirements.txt
alation-ai-agent-langchain~=0.1.1
langchain-community==0.3.23
langchain-openai==0.3.14
openai==1.76.2